requirements.txt
streamlit
PyPDF2~=2.0
google.generativeai
python-dotenv